added voice name transliteration.
